upgrade 3-extensions/dubbo-samples-nacos/dubbo-samples-nacos-conditionrouter to jdk17 and springboot3
