+++ 3-extensions/registry/dubbo-samples-nacos/dubbo-samples-nacos-conditionrouter/dubbo-samples-nacos-conditionrouter-interface/src/main/java/org/apache/dubbo/samples/governance/api/DemoService2.java
